Only use circuits when necessary (#9)
* Begin defining blazor.united.js

* Factoring out common parts of the server startup logic

* Only start a circuit when necessary
