+++ src/Components/Samples/BlazorUnitedApp/Shared/MainLayout.razor
@@ -33,6 +33,6 @@
         <a class="dismiss">X</a>
     </div>
 
-    <script src="_framework/blazor.server.js" suppress-error="BL9992"></script>
+    <script src="_framework/blazor.united.js" suppress-error="BL9992"></script>
 </body>
 </html>
