+++ src/Components/Server/src/Microsoft.AspNetCore.Components.Server.csproj
@@ -99,9 +99,12 @@
 
   <PropertyGroup>
     <BlazorServerJSFilename>blazor.server.js</BlazorServerJSFilename>
+    <BlazorUnitedJSFilename>blazor.united.js</BlazorUnitedJSFilename>
     <!-- Microsoft.AspNetCore.Components.Web.JS.npmproj always capitalizes the directory name. -->
     <BlazorServerJSFile Condition=" '$(Configuration)' == 'Debug' ">..\..\Web.JS\dist\Debug\$(BlazorServerJSFilename)</BlazorServerJSFile>
     <BlazorServerJSFile Condition=" '$(Configuration)' != 'Debug' ">..\..\Web.JS\dist\Release\$(BlazorServerJSFilename)</BlazorServerJSFile>
+    <BlazorUnitedJSFile Condition=" '$(Configuration)' == 'Debug' ">..\..\Web.JS\dist\Debug\$(BlazorUnitedJSFilename)</BlazorUnitedJSFile>
+    <BlazorUnitedJSFile Condition=" '$(Configuration)' != 'Debug' ">..\..\Web.JS\dist\Release\$(BlazorUnitedJSFilename)</BlazorUnitedJSFile>
   </PropertyGroup>
 
   <!-- blazor.server.js should exist after Microsoft.AspNetCore.Components.Web.JS.npmproj builds. Fall back if not. -->
@@ -115,7 +118,9 @@
   <Target Name="_AddEmbeddedBlazor" AfterTargets="_CheckBlazorServerJSPath">
     <ItemGroup>
       <EmbeddedResource Include="$(BlazorServerJSFile)" LogicalName="_framework/$(BlazorServerJSFilename)" />
+      <EmbeddedResource Include="$(BlazorUnitedJSFile)" LogicalName="_framework/$(BlazorUnitedJSFilename)" />
       <EmbeddedResource Include="$(BlazorServerJSFile).map" LogicalName="_framework/$(BlazorServerJSFilename).map" Condition="Exists('$(BlazorServerJSFile).map')" />
+      <EmbeddedResource Include="$(BlazorUnitedJSFile).map" LogicalName="_framework/$(BlazorUnitedJSFilename).map" Condition="Exists('$(BlazorUnitedJSFile).map')" />
     </ItemGroup>
   </Target>
 </Project>
