+++ src/Components/Web.JS/src/Platform/Circuits/UnitedStartOptions.ts
@@ -1,0 +1,8 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+import { CircuitStartOptions } from "./CircuitStartOptions";
+
+export interface UnitedStartOptions {
+  circuit: CircuitStartOptions;
+}
